requirements-dev.txt
# Include all production requirements
-r requirements.txt

# Development Tools
ipython>=8.14.0,<9.0.0
ipdb>=0.13.0,<1.0.0
django-extensions>=3.2.0,<4.0.0

# Testing and Quality Assurance
pytest-xdist>=3.3.0,<4.0.0  # Parallel test execution
pytest-mock>=3.11.0,<4.0.0
pytest-sugar>=0.9.0,<1.0.0
coverage>=7.3.0,<8.0.0
django-coverage-plugin>=3.1.0,<4.0.0

# Code Quality and Linting
pre-commit>=3.3.0,<4.0.0
bandit>=1.7.0,<2.0.0  # Security linting
safety>=2.3.0,<3.0.0  # Security vulnerability checker

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.2.0,<10.0.0

# Database Tools
django-extensions>=3.2.0,<4.0.0

# API Development
drf-spectacular>=0.26.0,<1.0.0  # OpenAPI/Swagger documentation
django-cors-headers>=4.2.0,<5.0.0

# Performance and Profiling
django-silk>=5.0.0,<6.0.0  # Request profiling
memory-profiler>=0.61.0,<1.0.0 